Add source and timeout flag to Cloud Functions deploy
Unfortunately in Cloud Functions, if you manually deploy via the Google Cloud Console, it overwrites the standard setting for the source flag. By adding the flag in the deploy script, we make sure it doesn't have an effect for the CI deployment because we reapply the source flag every time.

The timeout flag is for the system test runs and can be removed as soon as a continuously running instance for the CarML service is used. Otherwise the cold start prevents the tests from succeeding.